requirements.txt
click==7.1.2
fastapi==0.61.1
h11==0.10.0
psycopg2==2.8.6
pydantic==1.6.1
python-jose==3.3.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.12.0